Move from Lis\v2 to Membership
